+++ src/test/java/org/cthing/assertj/gradle/GradleAssertionsTest.java
@@ -9,6 +9,7 @@ import org.gradle.api.Project;
 import org.gradle.api.provider.Property;
 import org.gradle.api.provider.Provider;
 import org.gradle.testfixtures.ProjectBuilder;
+import org.jspecify.annotations.NonNull;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -38,10 +39,10 @@ public class GradleAssertionsTest {
 
     @Test
     public void testIsPresent() {
-        final Provider<String> provider1 = this.project.provider(() -> "test");
+        final Provider<@NonNull String> provider1 = this.project.provider(() -> "test");
         assertThat(provider1).isPresent();
 
-        final Property<String> property1 = this.project.getObjects().property(String.class).convention("value1");
+        final Property<@NonNull String> property1 = this.project.getObjects().property(String.class).convention("value1");
         assertThat(property1).isPresent();
     }
 }