+++ src/test/java/org/cthing/assertj/gradle/GradleFileCollectionAssertTest.java
@@ -13,6 +13,7 @@ import org.gradle.api.Project;
 import org.gradle.api.file.FileCollection;
 import org.gradle.api.provider.Provider;
 import org.gradle.testfixtures.ProjectBuilder;
+import org.jspecify.annotations.NonNull;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -27,7 +28,7 @@ public class GradleFileCollectionAssertTest {
     private FileCollection multiple;
     private FileCollection single;
     private FileCollection empty;
-    private Provider<FileCollection> provider;
+    private Provider<@NonNull FileCollection> provider;
 
     @BeforeEach
     public void setup() throws IOException {