+++ src/test/java/org/cthing/assertj/gradle/GradleProviderAssertTest.java
@@ -18,6 +18,7 @@ import org.gradle.api.provider.MapProperty;
 import org.gradle.api.provider.Property;
 import org.gradle.api.provider.Provider;
 import org.gradle.testfixtures.ProjectBuilder;
+import org.jspecify.annotations.NonNull;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -36,38 +37,39 @@ public class GradleProviderAssertTest {
 
     @Test
     public void testContains() {
-        final Provider<String> provider1 = this.project.provider(() -> "test");
+        final Provider<@NonNull String> provider1 = this.project.provider(() -> "test");
         assertThat(provider1).contains("test");
 
-        final Property<String> property1 = this.project.getObjects().property(String.class).convention("value1");
+        final Property<@NonNull String> property1 = this.project.getObjects().property(String.class).convention("value1");
         assertThat(property1).contains("value1");
         assertThatExceptionOfType(AssertionError.class)
                 .isThrownBy(() -> assertThat(property1).contains("value2"))
                 .withMessage("Expecting provider to contain 'value2' but was 'value1'");
 
-        final Property<String> property2 = this.project.getObjects().property(String.class);
+        final Property<@NonNull String> property2 = this.project.getObjects().property(String.class);
         assertThatExceptionOfType(AssertionError.class)
                 .isThrownBy(() -> assertThat(property2).contains("value1"))
                 .withMessage("Expecting provider to contain 'value1' but was empty");
 
-        final MapProperty<String, String> mapProperty = this.project.getObjects()
-                                                                    .mapProperty(String.class, String.class)
-                                                                    .convention(Map.of("a", "b", "c", "d"));
+        final MapProperty<@NonNull String, @NonNull String> mapProperty =
+                this.project.getObjects()
+                            .mapProperty(String.class, String.class)
+                            .convention(Map.of("a", "b", "c", "d"));
         assertThat(mapProperty).contains(Map.of("a", "b", "c", "d"));
     }
 
     @Test
     public void testContainsInstanceOf() {
-        final Provider<String> provider1 = this.project.provider(() -> "test");
+        final Provider<@NonNull String> provider1 = this.project.provider(() -> "test");
         assertThat(provider1).containsInstanceOf(String.class);
 
-        final Property<String> property1 = this.project.getObjects().property(String.class).convention("value1");
+        final Property<@NonNull String> property1 = this.project.getObjects().property(String.class).convention("value1");
         assertThat(property1).containsInstanceOf(String.class);
         assertThatExceptionOfType(AssertionError.class)
                 .isThrownBy(() -> assertThat(property1).containsInstanceOf(Integer.class))
                 .withMessage("Expecting 'DefaultProperty' to contain an instance of 'java.lang.Integer' but contained an instance of 'java.lang.String'");
 
-        final Property<String> property2 = this.project.getObjects().property(String.class);
+        final Property<@NonNull String> property2 = this.project.getObjects().property(String.class);
         assertThatExceptionOfType(AssertionError.class)
                 .isThrownBy(() -> assertThat(property2).containsInstanceOf(String.class))
                 .withMessage("Expecting 'property(java.lang.String, undefined)' to contain a value, but it was empty");
@@ -76,13 +78,13 @@ public class GradleProviderAssertTest {
     @Test
     public void testContainsSame() {
         final Integer value = 12345;
-        final Provider<Integer> provider1 = this.project.provider(() -> value);
+        final Provider<@NonNull Integer> provider1 = this.project.provider(() -> value);
         assertThat(provider1).containsSame(value);
         assertThatExceptionOfType(AssertionError.class)
                 .isThrownBy(() -> assertThat(provider1).containsSame(12345))
                 .withMessage("Expecting provider to contain value identical to '12345'");
 
-        final Property<Integer> property1 = this.project.getObjects().property(Integer.class);
+        final Property<@NonNull Integer> property1 = this.project.getObjects().property(Integer.class);
         assertThatExceptionOfType(AssertionError.class)
                 .isThrownBy(() -> assertThat(property1).containsSame(12345))
                 .withMessage("Expecting provider to contain '12345' but was empty");
@@ -90,35 +92,36 @@ public class GradleProviderAssertTest {
 
     @Test
     public void testHasValue() {
-        final Provider<String> provider1 = this.project.provider(() -> "test");
+        final Provider<@NonNull String> provider1 = this.project.provider(() -> "test");
         assertThat(provider1).hasValue("test");
 
-        final Property<String> property1 = this.project.getObjects().property(String.class).convention("value1");
+        final Property<@NonNull String> property1 = this.project.getObjects().property(String.class).convention("value1");
         assertThat(property1).hasValue("value1");
         assertThatExceptionOfType(AssertionError.class)
                 .isThrownBy(() -> assertThat(property1).hasValue("value2"))
                 .withMessage("Expecting provider to contain 'value2' but was 'value1'");
 
-        final Property<String> property2 = this.project.getObjects().property(String.class);
+        final Property<@NonNull String> property2 = this.project.getObjects().property(String.class);
         assertThatExceptionOfType(AssertionError.class)
                 .isThrownBy(() -> assertThat(property2).hasValue("value1"))
                 .withMessage("Expecting provider to contain 'value1' but was empty");
 
-        final MapProperty<String, String> mapProperty = this.project.getObjects()
-                                                                    .mapProperty(String.class, String.class)
-                                                                    .convention(Map.of("a", "b", "c", "d"));
+        final MapProperty<@NonNull String, @NonNull String> mapProperty =
+                this.project.getObjects()
+                            .mapProperty(String.class, String.class)
+                            .convention(Map.of("a", "b", "c", "d"));
         assertThat(mapProperty).hasValue(Map.of("a", "b", "c", "d"));
     }
 
     @Test
     public void testHasValueSatisfyingConsumer() {
-        final Provider<String> provider1 = this.project.provider(() -> "test");
+        final Provider<@NonNull String> provider1 = this.project.provider(() -> "test");
         assertThat(provider1).hasValueSatisfying(s -> Assertions.assertThat(s).isEqualTo("test"));
 
-        final Property<String> property1 = this.project.getObjects().property(String.class).convention("value1");
+        final Property<@NonNull String> property1 = this.project.getObjects().property(String.class).convention("value1");
         assertThat(property1).hasValueSatisfying(s -> Assertions.assertThat(s).isEqualTo("value1"));
 
-        final Property<String> property2 = this.project.getObjects().property(String.class);
+        final Property<@NonNull String> property2 = this.project.getObjects().property(String.class);
         assertThatExceptionOfType(AssertionError.class)
                 .isThrownBy(() -> assertThat(property2)
                         .hasValueSatisfying(s -> Assertions.assertThat(s).isEqualTo("test")))
@@ -129,13 +132,13 @@ public class GradleProviderAssertTest {
     public void testHasValueSatisfyingCondition() {
         final Condition<String> condition = new Condition<>("test"::equals, "testing");
 
-        final Provider<String> provider1 = this.project.provider(() -> "test");
+        final Provider<@NonNull String> provider1 = this.project.provider(() -> "test");
         assertThat(provider1).hasValueSatisfying(condition);
 
-        final Property<String> property1 = this.project.getObjects().property(String.class).convention("test");
+        final Property<@NonNull String> property1 = this.project.getObjects().property(String.class).convention("test");
         assertThat(property1).hasValueSatisfying(condition);
 
-        final Property<String> property2 = this.project.getObjects().property(String.class);
+        final Property<@NonNull String> property2 = this.project.getObjects().property(String.class);
         assertThatExceptionOfType(AssertionError.class)
                 .isThrownBy(() -> assertThat(property2).hasValueSatisfying(condition))
                 .withMessage("Expecting 'property(java.lang.String, undefined)' to contain a value, but it was empty");
@@ -143,13 +146,13 @@ public class GradleProviderAssertTest {
 
     @Test
     public void testIsPresent() {
-        final Provider<String> provider1 = this.project.provider(() -> "test");
+        final Provider<@NonNull String> provider1 = this.project.provider(() -> "test");
         assertThat(provider1).isPresent();
 
-        final Property<String> property1 = this.project.getObjects().property(String.class).convention("value1");
+        final Property<@NonNull String> property1 = this.project.getObjects().property(String.class).convention("value1");
         assertThat(property1).isPresent();
 
-        final Property<String> property2 = this.project.getObjects().property(String.class);
+        final Property<@NonNull String> property2 = this.project.getObjects().property(String.class);
         assertThatExceptionOfType(AssertionError.class)
                 .isThrownBy(() -> assertThat(property2).isPresent())
                 .withMessage("Expecting 'property(java.lang.String, undefined)' to contain a value, but it was empty");
@@ -157,29 +160,29 @@ public class GradleProviderAssertTest {
 
     @Test
     public void testIsEmpty() {
-        final Provider<String> provider1 = this.project.provider(() -> "test");
+        final Provider<@NonNull String> provider1 = this.project.provider(() -> "test");
         assertThatExceptionOfType(AssertionError.class)
                 .isThrownBy(() -> assertThat(provider1).isEmpty())
                 .withMessage("Expecting provider to be empty but contains 'provider(?)'");
 
-        final Property<String> property1 = this.project.getObjects().property(String.class).convention("value1");
+        final Property<@NonNull String> property1 = this.project.getObjects().property(String.class).convention("value1");
         assertThatExceptionOfType(AssertionError.class)
                 .isThrownBy(() -> assertThat(property1).isEmpty())
                 .withMessage("Expecting provider to be empty but contains 'property(java.lang.String, fixed(class java.lang.String, value1))'");
 
-        final Property<String> property2 = this.project.getObjects().property(String.class);
+        final Property<@NonNull String> property2 = this.project.getObjects().property(String.class);
          assertThat(property2).isEmpty();
     }
 
     @Test
     public void testGet() {
-        final Provider<String> provider1 = this.project.provider(() -> "test");
+        final Provider<@NonNull String> provider1 = this.project.provider(() -> "test");
         assertThat(provider1).get().isEqualTo("test");
 
-        final Property<String> property1 = this.project.getObjects().property(String.class).convention("value1");
+        final Property<@NonNull String> property1 = this.project.getObjects().property(String.class).convention("value1");
         assertThat(property1).get().isEqualTo("value1");
 
-        final Property<String> property2 = this.project.getObjects().property(String.class);
+        final Property<@NonNull String> property2 = this.project.getObjects().property(String.class);
         assertThatExceptionOfType(AssertionError.class)
                 .isThrownBy(() -> assertThat(property2).get().isNotNull())
                 .withMessage("Expecting 'property(java.lang.String, undefined)' to contain a value, but it was empty");
@@ -187,18 +190,18 @@ public class GradleProviderAssertTest {
 
     @Test
     public void testGetWithFactory() {
-        final Provider<String> provider1 = this.project.provider(() -> "test");
+        final Provider<@NonNull String> provider1 = this.project.provider(() -> "test");
         assertThat(provider1).get(InstanceOfAssertFactories.STRING).isEqualTo("test");
 
-        final Provider<List<String>> provider2 = this.project.provider(() -> List.of("a", "b"));
+        final Provider<@NonNull List<String>> provider2 = this.project.provider(() -> List.of("a", "b"));
         assertThat(provider2).get(InstanceOfAssertFactories.LIST).containsExactly("a", "b");
 
-        final Property<File> property1 = this.project.getObjects()
+        final Property<@NonNull File> property1 = this.project.getObjects()
                                                      .property(File.class)
                                                      .convention(new File("build/a.txt"));
         assertThat(property1).get(InstanceOfAssertFactories.FILE).hasName("a.txt");
 
-        final Property<String> property2 = this.project.getObjects().property(String.class);
+        final Property<@NonNull String> property2 = this.project.getObjects().property(String.class);
         assertThatExceptionOfType(AssertionError.class)
                 .isThrownBy(() -> assertThat(property2).get(InstanceOfAssertFactories.STRING).isNotNull())
                 .withMessage("Expecting 'property(java.lang.String, undefined)' to contain a value, but it was empty");
@@ -206,20 +209,20 @@ public class GradleProviderAssertTest {
 
     @Test
     public void testMap() {
-        final Provider<String> provider1 = this.project.provider(() -> "test1");
+        final Provider<@NonNull String> provider1 = this.project.provider(() -> "test1");
         assertThat(provider1).map(s -> "test2").contains("test2");
 
-        final Property<String> property1 = this.project.getObjects().property(String.class).convention("value1");
+        final Property<@NonNull String> property1 = this.project.getObjects().property(String.class).convention("value1");
         assertThat(property1).map(s -> "value2").contains("value2");
 
-        final Property<String> property2 = this.project.getObjects().property(String.class);
+        final Property<@NonNull String> property2 = this.project.getObjects().property(String.class);
         assertThat(property2).map(s -> null).isEmpty();
     }
 
     @Test
     public void testFlatMap() {
-        final Provider<String> provider1 = this.project.provider(() -> "test1");
-        final Provider<String> provider2 = this.project.provider(() -> "test2");
+        final Provider<@NonNull String> provider1 = this.project.provider(() -> "test1");
+        final Provider<@NonNull String> provider2 = this.project.provider(() -> "test2");
         assertThat(provider1).flatMap((Transformer<Provider<String>, String>)s -> provider2).contains("test2");
     }
 }